havenmind-basic/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai>=1.0.0
google-generativeai==0.3.2
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
pywhatkit==5.4
requests==2.31.0
reportlab==4.0.4
pyotp==2.9.0